Split dropbox_importer up into more source files
