+++ system_tests/src/dropbox_importer_tests.rs
@@ -6,8 +6,8 @@ use bytes::Bytes;
 use dogbox_tree::serialization::{DirectoryEntryKind, FileName};
 use dogbox_tree_editor::{FileCreationMode, OpenDirectory, OpenFile};
 use dropbox_importer::{
+    dropbox_api::RealDropboxApi,
     file_cache::{PersistableFileCache, PersistableFileCacheEntry, Sha256CacheKey},
-    RealDropboxApi,
 };
 use dropbox_sdk::{default_async_client::UserAuthDefaultClient, oauth2::Authorization};
 use futures::StreamExt;
